Refactor and update various files, add new services
- Remove unnecessary `using` directives and modify `Fruit` class in `Fruit.cs`
- Make `ICommand` interface public in `ICommand.cs`
- Change `Balance` property type from `int` to `double` in `User.cs`
- Refactor `SelectMenuExecuted.cs` to include `UserContext` and update methods
- Update `BonanzaSlot.cs` to reference `UserContext` and refactor methods
- Comment out `Execute` method for `SocketSlashCommand` in `ISlot.cs`
- Remove `CommandLoaderService.cs`
- Update `EventService.cs` to use `UserContext` and rename event method
- Replace `CommandLoaderService` with `LoaderService` in `SlashCommandService.cs`
- Add new `ButtonExecuted.cs` with an empty `ButtonExecuted` class
- Add new `LoaderService.cs` with methods for loading commands and slots

+++ new-discord-bot/Data/User.cs
@@ -9,7 +9,7 @@ namespace new_discord_bot.Data
 	public class User : IComparable<User>
 	{
 		public ulong Id { get; set; }
-		public int Balance { get; set; } = 100;
+		public double Balance { get; set; } = 100;
 		public int TotalSpins { get; set; } = 0;
 		public int TotalFeatures { get; set; } = 0;
 		public int HighestBalance { get; set; } = 0;
